# artifacts/api-server/src/routes/calls.ts
import { Router } from "express";
import { eq, desc, sql } from "drizzle-orm";
import { db, callsTable } from "@workspace/db";
import {
  ListCallsResponse,
  GetCallParams,
  GetCallResponse,
  UpdateCallParams,
  UpdateCallBody,
  UpdateCallResponse,
  DeleteCallParams,
  GetCallStatsResponse,
} from "@workspace/api-zod";

const router = Router();

function mapCall(c: typeof callsTable.$inferSelect) {
  return {
    ...c,
    transcript: Array.isArray(c.transcript) ? c.transcript : [],
    createdAt: c.createdAt.toISOString(),
  };
}

router.get("/calls/stats", async (req, res): Promise<void> => {
  const all = await db.select().from(callsTable);
  const now = new Date();
  const startOfDay = new Date(now.getFullYear(), now.getMonth(), now.getDate());
  const startOfWeek = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);

  const totalCalls = all.length;
  const bookedCalls = all.filter((c) => c.status === "booked").length;
  const missedCalls = all.filter((c) => c.status === "missed").length;
  const transferredCalls = all.filter((c) => c.status === "transferred").length;
  const totalDuration = all.reduce((sum, c) => sum + c.duration, 0);
  const averageDuration = totalCalls > 0 ? Math.round(totalDuration / totalCalls) : 0;
  const bookingRate = totalCalls > 0 ? Math.round((bookedCalls / totalCalls) * 100) / 100 : 0;
  const callsToday = all.filter((c) => c.createdAt >= startOfDay).length;
  const callsThisWeek = all.filter((c) => c.createdAt >= startOfWeek).length;

  res.json(GetCallStatsResponse.parse({
    totalCalls, bookedCalls, missedCalls, transferredCalls,
    averageDuration, bookingRate, callsToday, callsThisWeek,
  }));
});

router.get("/calls", async (req, res): Promise<void> => {
  let query = db.select().from(callsTable).orderBy(desc(callsTable.createdAt)).$dynamic();
  if (req.query.status && req.query.status !== "all") {
    query = query.where(eq(callsTable.status, req.query.status as string));
  }
  const limit = Number(req.query.limit) || 50;
  query = query.limit(limit);
  const rows = await query;
  res.json(ListCallsResponse.parse(rows.map(mapCall)));
});

router.get("/calls/:id", async (req, res): Promise<void> => {
  const params = GetCallParams.safeParse({ id: Number(req.params.id) });
  if (!params.success) { res.status(400).json({ error: "Invalid id" }); return; }
  const [row] = await db.select().from(callsTable).where(eq(callsTable.id, params.data.id));
  if (!row) { res.status(404).json({ error: "Not found" }); return; }
  res.json(GetCallResponse.parse(mapCall(row)));
});

router.patch("/calls/:id", async (req, res): Promise<void> => {
  const params = UpdateCallParams.safeParse({ id: Number(req.params.id) });
  if (!params.success) { res.status(400).json({ error: "Invalid id" }); return; }
  const parsed = UpdateCallBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: parsed.error.message }); return; }
  const [row] = await db
    .update(callsTable)
    .set(parsed.data)
    .where(eq(callsTable.id, params.data.id))
    .returning();
  if (!row) { res.status(404).json({ error: "Not found" }); return; }
  res.json(UpdateCallResponse.parse(mapCall(row)));
});

router.delete("/calls/:id", async (req, res): Promise<void> => {
  const params = DeleteCallParams.safeParse({ id: Number(req.params.id) });
  if (!params.success) { res.status(400).json({ error: "Invalid id" }); return; }
  await db.delete(callsTable).where(eq(callsTable.id, params.data.id));
  res.status(204).end();
});

export default router;