# lib/api-zod/src/generated/types/call.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BuildAI – Construction Company AI Assistant Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CallStatus } from './callStatus';

export interface Call {
  id: number;
  callerName: string;
  callerPhone: string;
  status: CallStatus;
  outcome: string;
  /** Duration in seconds */
  duration: number;
  assistantName: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  jobId?: number | null;
  /** @nullable */
  contactId?: number | null;
  reviewed?: boolean;
  createdAt: Date;
}